pyproject.toml
[project]
name = "ocs-docs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mkdocs-macros-plugin>=1.3.7",
    "mkdocs-material",
    "mkdocstrings[python]",
    "requests>=2.32.3",
]

[project.entry-points.'mkdocs.plugins']
api_docs = "ocs_docs.api_docs_plugin:ApiDocsBuildPlugin"

[build-system]
requires = ["uv_build>=0.8.10,<0.9.0"]
build-backend = "uv_build"
